backend/requirements.txt
amadeus
beautifulsoup4
openai
pandas
Requests
serpapi
streamlit
google-search-results
flask
flask_cors
gunicorn
python-dotenv
